requirements.txt
azure-cognitiveservices-speech==1.40.0